Adapting setuptools to allow local testing via docker images
Setuptools works well via docker, so long as we remove os.link, a change
that setuptools is tolerant to.  Hard-linking is an issue when crossing
some filesystems, so with os.link removed (at least for the purposes of
setuptools), we can install the package on docker as needed

This bug was fixed in python 2.7.9.
Branch: develop