Using standard GNU or BSD formats for checksum files
BSD and GNU have standard, allowed formats for files containing md5sums,
so no need to define our own.  The function to create this file will
write both styles (user can choose GNU or BSD style), and the assertion
for this will automatically detect which style is used in its assertion.

The particular power of this is that if the user wishes to use a
file generated by a native md5sum binary (such as GNU md5sum or BSD
md5), they can just provide that file ... it need not be generated by
out internal utility.
Branch: feature/geospatial-testing