Adding tests for asserting file digests
Two tests, one for asserting that we can test the assertions within a
different CWD, and the other for asserting a BSD-style checksum.

This test also revealed a bug in the implementation of
assert_checksums_equal that prevented the use of BSD checksum files.
I've corrected this and added a note to HISTORY about the change.

RE:#BITBUCKET-42
Branch: bugfix/TESTING-42-complete-test-coverage