Refactoring *-import test to avoid SyntaxWarning.
Rather than trying to actually use the *-import notation, the test for
this just asserts that everything in __all__ is available as a function
at the pygeoprocessing module level.  This has been tested against the
pygeoprocessing.__init__.py found in rev#BITBUCKET-159869233cea, so it should be a
reasonable guard against future failures of the *-import notation.

Fixes:#BITBUCKET-78
Branch: bugfix/IMPORTS-78-from-pygeoprocessing-import-star