Optimization for duplicate point geometries.
We sometimes have multiple points overlapping the same pixel.  If this happens
now, we will only determine watersheds for the first point, but copy over
the geometry to the new point (and fields from the original vector) at the end
of the delineation.  The result of this should be fewer passes over the raster.

RE:#BITBUCKET-130
Branch: feature/ROUTING-130-implement-d8-watershed-delineation-nested