Fragment joining looks like it works (though this needs testing).
The implementation in the fragment joining correctly implements
recursion, but lacks the avoided-recomputation that I'd prefer to have.
It's a starting point.  Takes about 196s to run on the roads sample
dataset and most of the time looks to be spent in `cascaded_union`.

RE:#BITBUCKET-130
Branch: feature/ROUTING-130-implement-d8-watershed-delineation-nested