Update issue #16 Support for GDAL 3
Adding support to handle GDAL 3. This effectively ends support for
GDAL 2. When using coordinate transformations now setting Axis
properties to be 'friendly' which means to expect lon,lat as opposed to
the now standard of lat,lon. This was recommonended in the migration
guide from GDAL.

Changing the values of a hardcoded test slightly as spatial references
are slightly different in GDAL3, so minor extent differences are
expected in some instances.