Issue #16 Fixing coodinatetransformation axis set
There is no need to check if the SRS is geographic or projected,
GDAL 3 appears to only swap when expected, like for lat/lon
to lon/lat for GIS friendly axis order.

Also removing the axis swapping in reproject_vector where the
geometries actually handle this internally.
Adding better comments.