Issue #16 Adding tests for Axis Ordering.
Adding tests to transform_bounding_box and reproject_vector functions
where osr.CreateCoordinateTransformation are handled. The tests for
transform bounding box use coordinates from a polygon made in QGIS
in the New England Coastal area for WGS84 and UTM 19N. The expected
results were varified in QGIS.

The test for reproject vector handle all possible transformation cases
in and out of WGS84 and Projected SRS. Specifically the latlon to latlon
test checks that the vectors are identical as well as the SRS is as
expected. To handle WGS84 I added a WGS84 reference to sampledata,
following the conventions for the other used SRS.