Squashing cherry-picked commits for clarity.
first cherry-pick to get updated actions which uses setup-miniconda
instead of setup-conda. This fixes the ubuntu testing issue when on GDAL
3. setup-miniconda is better documented and supported at this time.

Removing travis build config

Github actions seems to be doing a great job running our tests and
Travis is using docker which is currently running an outdated version of
GDAL and other dependencies.