Disabling diagnostic vector creation by default.
It turns out that on large geometries, the creation of the diagnostic
vector (which was always being created) was occupying the bulk of the
runtime: a new geometry is created for each seed, which requires calling
multiple python functions.

Rather than adding progress logging, I've added a flag to the watershed
delineation function itself to allow for a user to determine when the
diagnostic vectors should be created.  The flag defaults to False, so
the seed loop now happens in straight C/C++.

This fix completely avoids the need for any actual changes to progress
logging.

RE:#40