Defining known resample algorithms as a module constant.
GDAL's WarpOptions parsing function (``osgeo.gdal.WarpOptions``) has a
local dict defining a mapping of gdal.GRA_* constants to the appropriate
human-readable strings. The only ways to get at this mapping is to
either parse the source code with ``ast`` or to derive the
human-readable mappings by calling ``gdal.WarpOptions()`` multiple times
and getting the output.  Since we can't always rely on having the source
code available (like in pyinstaller), I went with the black-box method.
It appears to work fine and should be resilient to any changes to GDAL's
selection or naming of resample algorithms.

RE:#280