Creating TIFF kernels instead of GeoTIFFs.
I remember GDAL issuing a warning a long time ago if no SRS was defined,
but that no longer seems to be the case.  Both GDAL and QGIS seem to
have decided that the SRS doesn't matter any longer.  This commit
removes the spatial metadata information from kernel creation.

RE:#268